Extract image loader code & encapsulate logic
